fix in the data websocket the vital cpu temp form real to integer, fix the graphical element of cpu temperature in status.html

+++ src/workspace/src/hydroball_data_websocket/include/hydroball_data_websocket/hydroball_data_websocket.h
@@ -218,7 +218,7 @@ public:
         }
 	else{
             rapidjson::Value vitalsArray(rapidjson::Type::kArrayType);
-            rapidjson::Value cputemp(state.vitals.cputemp);
+            rapidjson::Value cputemp((int)state.vitals.cputemp);
             rapidjson::Value cpuload((int) state.vitals.cpuload);
             rapidjson::Value freeram((int) state.vitals.freeram);
             rapidjson::Value freehdd((int) state.vitals.freehdd);
